requirements.txt
# Evolving Battle AI - Core Dependencies
# Phase 1: Foundation & Architecture

# Core Python packages
numpy>=1.21.0
matplotlib>=3.5.0
pygame>=2.1.0

# Testing framework
pytest>=7.0.0
pytest-cov>=4.0.0

# Development tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Logging and configuration
pyyaml>=6.0
colorlog>=6.0.0

# Data handling
pandas>=1.4.0

# Future ML dependencies (will be added in Phase 2-3)
# tensorflow>=2.10.0  # Uncomment when needed
# torch>=1.12.0       # Alternative to TensorFlow
# stable-baselines3>=1.6.0  # For RL in Phase 3
# deap>=1.3.0         # For genetic algorithms
